Simplify integrations/lib/backoff
- move test-only code to a _test.go file
- consolidate code to fewer files
