+++ integrations/lib/backoff/backoff.go
@@ -18,9 +18,54 @@
 
 package backoff
 
-import "context"
+import (
+	"context"
+	"math/rand"
+	"time"
+
+	"github.com/gravitational/trace"
+	"github.com/jonboulle/clockwork"
+)
 
 // Backoff is an interface to some (exponential) backoff algorithm.
 type Backoff interface {
 	Do(context.Context) error
+}
+
+// decorr is a "decorrelated jitter" inspired by https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/.
+type decorr struct {
+	base  int64
+	cap   int64
+	mul   int64
+	sleep int64
+	clock clockwork.Clock
+}
+
+// NewDecorr initializes an algorithm.
+func NewDecorr(base, cap time.Duration, clock clockwork.Clock) Backoff {
+	return NewDecorrWithMul(base, cap, 3, clock)
+}
+
+// NewDecorrWithMul initializes a backoff algorithm with a given multiplier.
+func NewDecorrWithMul(base, cap time.Duration, mul int64, clock clockwork.Clock) Backoff {
+	return &decorr{
+		base:  int64(base),
+		cap:   int64(cap),
+		mul:   mul,
+		sleep: int64(base),
+		clock: clock,
+	}
+}
+
+func (backoff *decorr) Do(ctx context.Context) error {
+	backoff.sleep = backoff.base + rand.Int63n(backoff.sleep*backoff.mul-backoff.base)
+	if backoff.sleep > backoff.cap {
+		backoff.sleep = backoff.cap
+	}
+	select {
+	case <-backoff.clock.After(time.Duration(backoff.sleep)):
+		return nil
+	case <-ctx.Done():
+		return trace.Wrap(ctx.Err())
+	}
 }
